backend/requirements.txt
# FastAPI Backend Requirements - Compatible with Python 3.12
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
alembic==1.14.0

# Image Processing
Pillow==11.0.0
opencv-python==4.10.0.84

# AI/ML - CLIP Model
torch==2.5.1
torchvision==0.20.1
transformers==4.46.3
ftfy==6.3.1
regex==2024.11.6

# EXIF Metadata
exifread==3.0.0
piexif==1.1.3

# File handling
aiofiles==24.1.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.3.0

# Testing
pytest==8.3.4
httpx==0.28.1
